removing census layer using the vocdoni-node censusdb package directly
